+++ plugins/polio/api/orgunits_per_campaigns.py
@@ -1,0 +1,74 @@
+import csv
+
+from django.db.models import TextField, UUIDField, Value
+from django.http import HttpResponse, JsonResponse
+from rest_framework import viewsets
+
+from iaso.api.common import CONTENT_TYPE_CSV
+from iaso.models import OrgUnit
+from plugins.polio.models import Campaign
+
+
+def org_unit_as_array(o):
+    res = [o.campaign_id, o.campaign_obr, o.id, o.name, o.org_unit_type.name]
+
+    parent = o
+    for i in range(4):
+        if parent:
+            parent = parent.parent
+        if parent:
+            res.extend([parent.id, parent.name])
+        else:
+            res.extend((None, None))
+    return res
+
+
+class OrgUnitsPerCampaignViewset(viewsets.ViewSet):
+    def list(self, request):
+        org_unit_type = request.GET.get("org_unit_type_id", None)
+        as_csv = request.GET.get("format", None) == "csv"
+        campaigns = Campaign.objects.all()
+        queryset = OrgUnit.objects.none()
+
+        for campaign in campaigns:
+            districts = campaign.get_all_districts()
+
+            if districts:
+                all_facilities = OrgUnit.objects.hierarchy(districts)
+                if org_unit_type:
+                    all_facilities = all_facilities.filter(org_unit_type_id=org_unit_type)
+                all_facilities = (
+                    all_facilities.prefetch_related("parent")
+                    .prefetch_related("parent__parent")
+                    .prefetch_related("parent__parent__parent")
+                    .prefetch_related("parent__parent__parent__parent")
+                )
+                all_facilities = all_facilities.annotate(campaign_id=Value(campaign.id, UUIDField()))
+                all_facilities = all_facilities.annotate(campaign_obr=Value(campaign.obr_name, TextField()))
+                queryset = queryset.union(all_facilities)
+
+        headers = [
+            "campaign_id",
+            "campaign_obr",
+            "org_unit_id",
+            "org_unit_name",
+            "type",
+            "parent1_id",
+            "parent1_name",
+            "parent2_id",
+            "parent2_name",
+            "parent3_id",
+            "parent3_name",
+            "parent4_id",
+            "parent4_name",
+        ]
+        res = [headers]
+        res.extend([org_unit_as_array(o) for o in queryset])
+        if as_csv:
+            response = HttpResponse(content_type=CONTENT_TYPE_CSV)
+            writer = csv.writer(response)
+            for item in res:
+                writer.writerow(item)
+            return response
+        else:
+            return JsonResponse(res, safe=False)
