+++ plugins/polio/api/preparedness_dashboard.py
@@ -1,0 +1,20 @@
+from rest_framework import viewsets
+from rest_framework.response import Response
+
+from plugins.polio.models import Campaign
+from plugins.polio.preparedness.summary import get_or_set_preparedness_cache_for_round
+
+
+class PreparednessDashboardViewSet(viewsets.ViewSet):
+    def list(self, request):
+        r = []
+        qs = Campaign.objects.all().prefetch_related("rounds")
+        if request.query_params.get("campaign"):
+            qs = qs.filter(obr_name=request.query_params.get("campaign"))
+
+        for c in qs:
+            for round in c.rounds.all():
+                p = get_or_set_preparedness_cache_for_round(c, round)
+                if p:
+                    r.append(p)
+        return Response(r)
